+++ lpu/metrics/__init__.py
@@ -2,6 +2,7 @@
 
 __all__ = [
   'bleu',
+  'ranking',
   'ribes',
 ]